added restaurant pages+rest.list
