--- src/components/Contexts.js
@@ -1,3 +1,0 @@
-import React from "react";
-
-export const ContextDemo = React.createContext();
