--- src/components/GetRequests.module.css
@@ -1,7 +1,0 @@
-.itemContainer {
-    display:flex;
-}
-.itemContainer> * {
-    margin-right: 10px;
-    font-size: 17px;
-}
